random_chat/requirements.txt
Django==4.1.1
djangorestframework==3.14.0
channels[daphne]==4.0.0
channels-redis==4.0.0
asgiref==3.5.2
django-cors-headers==3.14.0
python-decouple==3.7
